requirements.txt
requests
beautifulsoup4
selenium
pytesseract
opencv-python
scapy
python-dotenv
sqlite3
beautifulsoup4
python-dotenv

